requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.5
billiard==4.2.1
black==25.1.0
cachetools==5.5.2
celery==5.3.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
cron_descriptor==2.0.6
cryptography==43.0.1
defusedxml==0.7.1
dj-database-url==3.0.1
dj-rest-auth==5.0.2
Django==5.2.6
django-allauth==0.57.0
django-celery-beat==2.8.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-csp==4.0
django-db-connection-pool==1.2.6
django-debug-toolbar==6.0.0
django-jazzmin==3.0.1
django-ratelimit==4.1.0
django-redis==6.0.0
django-timezone-field==7.1
django-uuid7==0.1.1
django_celery_results==2.6.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.28.0
Faker==37.6.0
gemini-cli==0.2.5
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.62.3
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
markdown-it-py==3.0.0
mdurl==0.1.2
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==4.25.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.8.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparams==6.2.0
sqlparse==0.5.3
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid6==2025.0.1
uuid7==0.1.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0